Introduce quadratic time SPH
The major change in this commit is in `simulation.rs` and `particle.rs` where a quadratic time SPH is implemented. This change causes some smaller changes in other files. In addition, this commit improves the observer with an idealised system radius, places the constants in their own file, and does some other small architectural changes.